Centralize stripe network client provider.
