+++ crypto-onramp/src/main/java/com/stripe/android/crypto/onramp/di/OnrampModule.kt
@@ -7,12 +7,10 @@ import com.stripe.android.BuildConfig
 import com.stripe.android.Stripe
 import com.stripe.android.common.di.MobileSessionIdModule
 import com.stripe.android.core.ApiVersion
-import com.stripe.android.core.Logger
 import com.stripe.android.core.injection.ENABLE_LOGGING
-import com.stripe.android.core.injection.IOContext
 import com.stripe.android.core.injection.PUBLISHABLE_KEY
 import com.stripe.android.core.injection.STRIPE_ACCOUNT_ID
-import com.stripe.android.core.networking.DefaultStripeNetworkClient
+import com.stripe.android.core.injection.StripeNetworkClientModule
 import com.stripe.android.core.networking.StripeNetworkClient
 import com.stripe.android.crypto.onramp.analytics.OnrampAnalyticsService
 import com.stripe.android.crypto.onramp.analytics.OnrampAnalyticsServiceImpl
@@ -26,10 +24,9 @@ import dagger.Module
 import dagger.Provides
 import javax.inject.Named
 import javax.inject.Singleton
-import kotlin.coroutines.CoroutineContext
 
 @Module(
-    includes = [MobileSessionIdModule::class, PaymentConfigurationModule::class],
+    includes = [MobileSessionIdModule::class, PaymentConfigurationModule::class, StripeNetworkClientModule::class],
     subcomponents = [OnrampPresenterComponent::class]
 )
 internal class OnrampModule {
@@ -39,15 +36,6 @@ internal class OnrampModule {
 
     @Provides
     fun provideAppContext(application: Application): Context = application.applicationContext
-
-    @Provides
-    fun provideStripeNetworkClient(
-        @IOContext context: CoroutineContext,
-        logger: Logger
-    ): StripeNetworkClient = DefaultStripeNetworkClient(
-        workContext = context,
-        logger = logger
-    )
 
     @Provides
     fun provideRequestSurface(): RequestSurface = RequestSurface.CryptoOnramp
