+++ identity/src/main/java/com/stripe/android/identity/injection/IdentityVerificationSheetComponent.kt
@@ -1,7 +1,9 @@
 package com.stripe.android.identity.injection
 
 import android.content.Context
+import com.stripe.android.core.injection.CoreCommonModule
 import com.stripe.android.core.injection.CoroutineContextModule
+import com.stripe.android.core.injection.StripeNetworkClientModule
 import com.stripe.android.identity.IdentityActivity
 import dagger.BindsInstance
 import dagger.Component
@@ -11,7 +13,9 @@ import javax.inject.Singleton
 @Component(
     modules = [
         IdentityCommonModule::class,
-        CoroutineContextModule::class
+        CoreCommonModule::class,
+        CoroutineContextModule::class,
+        StripeNetworkClientModule::class,
     ]
 )
 internal interface IdentityVerificationSheetComponent {
